backend/requirements.txt
fastapi
uvicorn
numpy
pandas
pillow
tensorflow-cpu
python-multipart
pydantic
joblib
